requirements.txt
fastmcp==2.13.3
openai>=1.0.0
python-dotenv

